Created level spawn point and goal functionality
Deleted level scriptable objects and moved all data to the level script itself.
Levels now hold data on where the player will spawn, the goal tile of the player, and the next level to be loaded.

Also deleted Enemy Data scriptable object and moved data to the Enemy script itself.

Touched up test levels so the goals are clear