Refactored scripts
Refactored LevelManager to have the CurrentLevel be static across the program.

Refactored TargetFollower to not use the tag finding behaviour. Now if you set the target of the follower to the object containing the TargetFollower script itself, it will create a personal target for the object to follow and attach it to the "Target Container" tagged object.

Refactored and cleaned up code for MovingEntity and Player.

Moved LevelManager script to the Grid